requirements.txt
# Consolidated requirements file for the SoftCap project
# This single file contains both runtime and development dependencies.
# It's intended for single-user research workflows. If you later want a
# minimal runtime-only set, we can split this again.

# Core ML Framework
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0

# Data Analysis & Statistics
pandas>=2.0.0
scipy>=1.11.0
scikit-learn>=1.3.0
statsmodels>=0.14.0

# Visualization & Plotting
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
bokeh>=3.2.0
pillow>=10.0.0
opencv-python>=4.8.0
imageio>=2.31.0

# Progress & UI
tqdm>=4.65.0
ipywidgets>=8.0.0

# Configuration & I/O
pyyaml>=6.0
json5>=0.9.0
toml>=0.10.0

# Synthetic Dataset Generation & Research Utilities
noise>=1.2.2
perlin-noise>=1.12
scikit-image>=0.21.0
networkx>=3.1.0
sympy>=1.12.0

# Testing, Linting & Static Analysis (Development)
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-xdist>=3.3.0
black>=23.7.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.5.0
pre-commit>=3.3.0

# Documentation & Docs-theme
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0

# Development Tools & Profiling
ipdb>=0.13.0
memory-profiler>=0.61.0
line-profiler>=4.1.0

# App frameworks & Advanced Visualization
dash>=2.14.0
streamlit>=1.25.0
gradio>=3.40.0

# Experiment tooling
wandb>=0.15.0
tensorboard>=2.14.0
mlflow>=2.5.0

# GPU Monitoring & Profiling
gpustat>=1.1.0
nvidia-ml-py>=12.535.0
psutil>=5.9.0

# Performance & Optimization
numba>=0.57.0
cython>=3.0.0

# Packaging / runtime helpers
# `pkg_resources` is provided by setuptools; added explicitly for reproducible environments
setuptools>=65.0.0

# NLP Scaling Benchmark (DistilBERT / GLUE) — install when running benchmarks/nlp/
transformers>=4.36.0
datasets>=2.16.0
accelerate>=0.25.0
evaluate>=0.4.0

# End of consolidated requirements